+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.4
+
+* Context issue resolved.
+
 ## 0.0.1
 
-* TODO: its a initial release only for android to connect helix watch usingg flutter to android refer to example to know how to use.
+* its a initial release only for android to connect helix watch usingg flutter to android refer to example to know how to use.